Window Size Linux

+++ src/main.rs
@@ -16,7 +16,7 @@ fn main() -> Result<(), String> {
     }
     tui_terminal.clear_screen();
     loop {
-        let (a, b): (u32, u32) = tui_terminal
+        let (a, b): (u16, u16) = tui_terminal
             .get_teminal_size()
             .ok_or("Failed To get Size")?;
         tui_terminal.println(format!("{}, {}", a, b));
